# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Cycling Fantasy Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface UserProfile {
  id: number;
  replitId: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  teamName: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 50
     */
  teamName: string;
}

export type RiderCurrentJerseysItem = typeof RiderCurrentJerseysItem[keyof typeof RiderCurrentJerseysItem];


export const RiderCurrentJerseysItem = {
  yellow: 'yellow',
  green: 'green',
  polkadot: 'polkadot',
  white: 'white',
} as const;

export interface Rider {
  id: number;
  name: string;
  proTeam: string;
  nationality: string;
  oddsDecimal: number;
  /** Human-readable odds e.g. "1/4" or "10/1" */
  oddsLabel: string;
  /** @nullable */
  photoUrl?: string | null;
  isActive: boolean;
  dnf: boolean;
  currentJerseys: RiderCurrentJerseysItem[];
  /**
     * ProCyclingStats rider slug for scraping
     * @nullable
     */
  pcsSlug?: string | null;
}

export interface RiderUpdate {
  oddsDecimal?: number;
  oddsLabel?: string;
  isActive?: boolean;
  dnf?: boolean;
}

export type StageStageType = typeof StageStageType[keyof typeof StageStageType];


export const StageStageType = {
  flat: 'flat',
  hilly: 'hilly',
  mountain: 'mountain',
  time_trial: 'time_trial',
  rest: 'rest',
} as const;

export type StageStatus = typeof StageStatus[keyof typeof StageStatus];


export const StageStatus = {
  upcoming: 'upcoming',
  transfer_closed: 'transfer_closed',
  live: 'live',
  completed: 'completed',
} as const;

export interface Stage {
  id: number;
  stageNumber: number;
  name: string;
  startCity: string;
  endCity: string;
  date: string;
  stageType: StageStageType;
  status: StageStatus;
  /** @nullable */
  transferDeadline: string | null;
  /** @nullable */
  pcsUrl?: string | null;
}

export interface CurrentStageResponse {
  stage: Stage;
  transferWindowOpen: boolean;
  /** @nullable */
  minutesUntilClose: number | null;
}

export type StageResultPointsBreakdown = {
  stage: number;
  jerseys: number;
  kom: number;
  sprint: number;
  combative: number;
  penalty: number;
};

export interface StageResult {
  id: number;
  stageId: number;
  riderId: number;
  riderName: string;
  proTeam: string;
  /** @nullable */
  position: number | null;
  dnf: boolean;
  fantasyPoints: number;
  pointsBreakdown?: StageResultPointsBreakdown;
}

export interface StageDetail {
  stage: Stage;
  results: StageResult[];
}

export interface RiderSummary {
  id: number;
  name: string;
  proTeam: string;
  oddsDecimal: number;
  oddsLabel: string;
  /** @nullable */
  photoUrl?: string | null;
  currentJerseys: string[];
  dnf: boolean;
}

export interface MyTeam {
  riders: RiderSummary[];
  /** @nullable */
  captainRiderId: number | null;
  transferWindowOpen: boolean;
  totalPoints?: number;
}

export interface TeamUpdate {
  /**
     * @minItems 1
     * @maxItems 8
     */
  riderIds: number[];
  captainRiderId: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  teamName: string;
  totalPoints: number;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  myRank?: boolean;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  myEntry: LeaderboardEntry | null;
}

export type StagePointsEntryBreakdown = {
  stage: number;
  jerseys: number;
  kom: number;
  sprint: number;
  combative: number;
  penalty: number;
};

export interface StagePointsEntry {
  stageId: number;
  stageNumber: number;
  stageName: string;
  riderId?: number;
  riderName: string;
  isCaptain: boolean;
  oddsDecimal?: number;
  basePoints: number;
  multiplier: number;
  totalPoints: number;
  breakdown: StagePointsEntryBreakdown;
}

export type MyPointsSummaryStageBreakdownItem = {
  stageId: number;
  stageNumber: number;
  stageName: string;
  points: number;
};

export interface MyPointsSummary {
  totalPoints: number;
  /** @nullable */
  teamName: string | null;
  stageBreakdown: MyPointsSummaryStageBreakdownItem[];
}

export interface StagePointsDetail {
  stageId: number;
  stageNumber: number;
  stageName: string;
  totalPoints: number;
  riderPoints: StagePointsEntry[];
}

export interface AdminStage {
  id: number;
  stageNumber: number;
  name: string;
  startCity?: string;
  endCity?: string;
  date: string;
  stageType?: string;
  status: string;
  /** @nullable */
  transferDeadline: string | null;
  /** @nullable */
  pcsUrl?: string | null;
  resultsProcessed: boolean;
}

export type AdminStageUpdateStatus = typeof AdminStageUpdateStatus[keyof typeof AdminStageUpdateStatus];


export const AdminStageUpdateStatus = {
  upcoming: 'upcoming',
  transfer_closed: 'transfer_closed',
  live: 'live',
  completed: 'completed',
} as const;

export interface AdminStageUpdate {
  date?: string;
  status?: AdminStageUpdateStatus;
  transferDeadline?: string;
  pcsUrl?: string;
}

export interface ProcessStageResult {
  success: boolean;
  ridersProcessed: number;
  message: string;
  errors?: string[];
}

export interface SyncResult {
  success: boolean;
  ridersFound: number;
  message: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type ListRidersParams = {
search?: string;
team?: string;
};

export type GetLeaderboardParams = {
limit?: number;
offset?: number;
};