# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cycling Fantasy Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [team]
      summary: Get the current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
    put:
      operationId: updateProfile
      tags: [team]
      summary: Update the current user's team name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated

  # Riders
  /riders:
    get:
      operationId: listRiders
      tags: [riders]
      summary: List all TDF riders
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: team
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"
  /riders/{id}:
    get:
      operationId: getRider
      tags: [riders]
      summary: Get a single rider by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rider detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
        "404":
          description: Rider not found

  # Stages
...
  /points/me:
    get:
      operationId: getMyPoints
      tags: [points]
      summary: Get my total points and per-stage breakdown
      responses:
        "200":
          description: My points summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyPointsSummary"
        "401":
          description: Not authenticated
  /points/stage/{stageId}:
    get:
      operationId: getStagePoints
      tags: [points]
      summary: Get my detailed points breakdown for a specific stage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stage points detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StagePointsDetail"
        "401":
          description: Not authenticated
        "404":
          description: Stage not found

  # Admin
  /admin/stages:
    get:
      operationId: adminListStages
      tags: [admin]
      summary: List all stages for admin management
      responses:
        "200":
          description: Admin stage list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStage"
        "403":
          description: Forbidden
  /admin/stages/{id}/process:
    post:
      operationId: adminProcessStage
      tags: [admin]
      summary: Scrape stage results and calculate points
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stage processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessStageResult"
        "400":
          description: Stage already processed or not completed
        "403":
          description: Forbidden
  /admin/stages/{id}:
    put:
      operationId: adminUpdateStage
      tags: [admin]
      summary: Update stage details (date, deadline, status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStageUpdate"
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStage"
        "403":
          description: Forbidden
  /admin/riders/{id}:
    put:
      operationId: adminUpdateRider
      tags: [admin]
      summary: Update rider odds or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiderUpdate"
      responses:
        "200":
          description: Updated rider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"
        "403":
          description: Forbidden
  /admin/sync:
    post:
      operationId: adminSyncRiders
      tags: [admin]
      summary: Sync rider data from ProCyclingStats
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "403":
          description: Forbidden
components:
...
  schemas:
...
    UserProfile:
      type: object
      required: [id, replitId, email, firstName, teamName, isAdmin, createdAt]
      properties:
        id:
          type: integer
        replitId:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      required: [teamName]
      properties:
        teamName:
          type: string
          minLength: 1
          maxLength: 50
    Rider:
      type: object
      required: [id, name, proTeam, nationality, oddsDecimal, oddsLabel, isActive, currentJerseys, dnf]
      properties:
        id:
          type: integer
        name:
          type: string
        proTeam:
          type: string
        nationality:
          type: string
        oddsDecimal:
          type: number
        oddsLabel:
          type: string
          description: Human-readable odds e.g. "1/4" or "10/1"
        photoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        dnf:
          type: boolean
        currentJerseys:
          type: array
          items:
            type: string
            enum: [yellow, green, polkadot, white]
        pcsSlug:
          type: ["string", "null"]
          description: ProCyclingStats rider slug for scraping
    RiderUpdate:
      type: object
      properties:
        oddsDecimal:
          type: number
        oddsLabel:
          type: string
        isActive:
          type: boolean
        dnf:
          type: boolean
...
    StagePointsEntry:
      type: object
      required: [stageId, stageNumber, stageName, riderName, isCaptain, basePoints, multiplier, totalPoints, breakdown]
      properties:
        stageId:
          type: integer
        stageNumber:
          type: integer
        stageName:
          type: string
        riderId:
          type: integer
        riderName:
          type: string
        isCaptain:
          type: boolean
        oddsDecimal:
          type: number
        basePoints:
          type: number
        multiplier:
          type: number
        totalPoints:
          type: number
        breakdown:
          type: object
          properties:
            stage:
              type: number
            jerseys:
              type: number
            kom:
              type: number
            sprint:
              type: number
            combative:
              type: number
            penalty:
              type: number
          required: [stage, jerseys, kom, sprint, combative, penalty]
    MyPointsSummary:
      type: object
      required: [totalPoints, stageBreakdown, teamName]
      properties:
        totalPoints:
          type: number
        teamName:
          type: ["string", "null"]
        stageBreakdown:
          type: array
          items:
            type: object
            required: [stageId, stageNumber, stageName, points]
            properties:
              stageId:
                type: integer
              stageNumber:
                type: integer
              stageName:
                type: string
              points:
                type: number
    StagePointsDetail:
      type: object
      required: [stageId, stageNumber, stageName, totalPoints, riderPoints]
      properties:
        stageId:
          type: integer
        stageNumber:
          type: integer
        stageName:
          type: string
        totalPoints:
          type: number
        riderPoints:
          type: array
          items:
            $ref: "#/components/schemas/StagePointsEntry"
    AdminStage:
      type: object
      required: [id, stageNumber, name, date, status, transferDeadline, resultsProcessed]
      properties:
        id:
          type: integer
        stageNumber:
          type: integer
        name:
          type: string
        startCity:
          type: string
        endCity:
          type: string
        date:
          type: string
          format: date
        stageType:
          type: string
        status:
          type: string
        transferDeadline:
          type: ["string", "null"]
          format: date-time
        pcsUrl:
          type: ["string", "null"]
        resultsProcessed:
          type: boolean
    AdminStageUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        status:
          type: string
          enum: [upcoming, transfer_closed, live, completed]
        transferDeadline:
          type: string
          format: date-time
        pcsUrl:
          type: string
    ProcessStageResult:
      type: object
      required: [success, ridersProcessed, message]
      properties:
        success:
          type: boolean
        ridersProcessed:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    SyncResult:
      type: object
      required: [success, ridersFound, message]
      properties:
        success:
          type: boolean
        ridersFound:
          type: integer
        message:
          type: string